Added DataResolver for handling data with ContentProvider
- Added Query Builder for faster data selection with content provider
- Minor bug fixes and improvements
