--- app/src/main/java/com/oogbox/support/orm/annotation/DataModel.java
@@ -1,12 +1,0 @@
-package com.oogbox.support.orm.annotation;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Retention(RetentionPolicy.RUNTIME)
-@Target(ElementType.TYPE)
-public @interface DataModel {
-    String value();
-}
