+++ app/src/main/java/com/oogbox/support/orm/core/annotation/DataModel.java
@@ -1,0 +1,12 @@
+package com.oogbox.support.orm.core.annotation;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Retention(RetentionPolicy.RUNTIME)
+@Target(ElementType.TYPE)
+public @interface DataModel {
+    String value();
+}
