+++ app/src/main/java/com/oogbox/support/orm/core/data/RelationOperation.java
@@ -1,0 +1,5 @@
+package com.oogbox.support.orm.core.data;
+
+public enum RelationOperation {
+    APPEND, REPLACE, REMOVE, SET_NULL
+}
