+++ app/src/main/java/com/oogbox/support/orm/core/listeners/MobileORMConfigListener.java
@@ -1,0 +1,14 @@
+package com.oogbox.support.orm.core.listeners;
+
+import android.content.Context;
+
+import com.oogbox.support.orm.BaseModel;
+
+import java.util.List;
+
+public interface MobileORMConfigListener {
+    List<BaseModel> getModels(Context context);
+
+    String getDatabaseName(Context context);
+
+}
