+++ app/src/main/java/com/oogbox/support/orm/core/types/OText.java
@@ -1,0 +1,15 @@
+package com.oogbox.support.orm.core.types;
+
+import com.oogbox.support.orm.core.types.helper.OFieldType;
+
+public class OText extends OFieldType<OText> {
+
+    public OText(String label) {
+        super(label);
+    }
+
+    @Override
+    public String fieldTypeString() {
+        return "TEXT";
+    }
+}
