--- app/src/main/java/com/oogbox/support/orm/data/RelationOperation.java
@@ -1,5 +1,0 @@
-package com.oogbox.support.orm.data;
-
-public enum RelationOperation {
-    APPEND, REPLACE, REMOVE, SET_NULL
-}
