--- app/src/main/java/com/oogbox/support/orm/types/OBlob.java
@@ -1,15 +1,0 @@
-package com.oogbox.support.orm.types;
-
-import com.oogbox.support.orm.types.helper.OFieldType;
-
-public class OBlob extends OFieldType<OBlob> {
-
-    public OBlob(String label) {
-        super(label);
-    }
-
-    @Override
-    public String fieldTypeString() {
-        return "BLOB";
-    }
-}
