--- app/src/main/java/com/oogbox/support/orm/types/ODate.java
@@ -1,15 +1,0 @@
-package com.oogbox.support.orm.types;
-
-import com.oogbox.support.orm.types.helper.OFieldType;
-
-public class ODate extends OFieldType<ODate> {
-
-    public ODate(String label) {
-        super(label);
-    }
-
-    @Override
-    public String fieldTypeString() {
-        return "DATE";
-    }
-}
