--- app/src/main/java/com/oogbox/support/orm/types/ODateTime.java
@@ -1,15 +1,0 @@
-package com.oogbox.support.orm.types;
-
-import com.oogbox.support.orm.types.helper.OFieldType;
-
-public class ODateTime extends OFieldType<ODateTime> {
-
-    public ODateTime(String label) {
-        super(label);
-    }
-
-    @Override
-    public String fieldTypeString() {
-        return "DATETIME";
-    }
-}
