--- app/src/main/java/com/oogbox/support/orm/types/OEnum.java
@@ -1,15 +1,0 @@
-package com.oogbox.support.orm.types;
-
-import com.oogbox.support.orm.types.helper.OFieldType;
-
-public class OEnum extends OFieldType<OEnum> {
-
-    public OEnum(String label) {
-        super(label);
-    }
-
-    @Override
-    public String fieldTypeString() {
-        return "VARCHAR";
-    }
-}
