+++ app/src/main/java/com/oogbox/support/orm/utils/ObjectByteUtils.java
@@ -1,0 +1,23 @@
+package com.oogbox.support.orm.utils;
+
+import java.io.ByteArrayInputStream;
+import java.io.ByteArrayOutputStream;
+import java.io.IOException;
+import java.io.ObjectInputStream;
+import java.io.ObjectOutputStream;
+
+public class ObjectByteUtils {
+
+    public static byte[] objectToByte(Object obj) throws IOException {
+        ByteArrayOutputStream outputStream = new ByteArrayOutputStream();
+        ObjectOutputStream os = new ObjectOutputStream(outputStream);
+        os.writeObject(obj);
+        return outputStream.toByteArray();
+    }
+
+    public static Object byteToObject(byte[] data) throws IOException, ClassNotFoundException {
+        ByteArrayInputStream inputStream = new ByteArrayInputStream(data);
+        ObjectInputStream is = new ObjectInputStream(inputStream);
+        return is.readObject();
+    }
+}
